+++ packages/web/app/components/PatientPrinting/modals/CovidClearanceCertificateModal.jsx
@@ -10,7 +10,7 @@ import { EmailButton } from '../../Email/EmailButton';
 import { useCertificate } from '../../../utils/useCertificate';
 import { usePatientAdditionalDataQuery } from '../../../api/queries';
 
-import { PDFViewer, printPDF } from '../PDFViewer';
+import { PDFLoader, printPDF } from '../PDFLoader';
 import { useCovidLabTestQuery } from '../../../api/queries/useCovidLabTestsQuery';
 
 export const CovidClearanceCertificateModal = React.memo(({ patient }) => {
@@ -55,7 +55,7 @@ export const CovidClearanceCertificateModal = React.memo(({ patient }) => {
       onPrint={() => printPDF('clearance-certificate')}
       additionalActions={<EmailButton onEmail={createCovidTestCertNotification} />}
     >
-      <PDFViewer isLoading={isLoading} id="clearance-certificate">
+      <PDFLoader isLoading={isLoading} id="clearance-certificate">
         <CovidLabCertificate
           patient={patientData}
           labs={labTestsResponse?.data}
@@ -66,7 +66,7 @@ export const CovidClearanceCertificateModal = React.memo(({ patient }) => {
           printedBy={printedBy}
           certType={CertificateTypes.clearance}
         />
-      </PDFViewer>
+      </PDFLoader>
     </Modal>
   );
 });
