+++ packages/web/app/components/PatientPrinting/modals/CovidVaccineCertificateModal.jsx
@@ -11,7 +11,7 @@ import { useCertificate } from '../../../utils/useCertificate';
 import { useLocalisation } from '../../../contexts/Localisation';
 import { useAdministeredVaccines, usePatientAdditionalDataQuery } from '../../../api/queries';
 
-import { PDFViewer, printPDF } from '../PDFViewer';
+import { PDFLoader, printPDF } from '../PDFLoader';
 
 export const CovidVaccineCertificateModal = React.memo(({ open, onClose, patient }) => {
   const api = useApi();
@@ -57,7 +57,7 @@ export const CovidVaccineCertificateModal = React.memo(({ open, onClose, patient
       onPrint={() => printPDF('covid-vaccine-certificate')}
       additionalActions={<EmailButton onEmail={createCovidVaccineCertificateNotification} />}
     >
-      <PDFViewer isLoading={isLoading} id="covid-vaccine-certificate">
+      <PDFLoader isLoading={isLoading} id="covid-vaccine-certificate">
         <CovidVaccineCertificate
           patient={patientData}
           vaccinations={vaccinations}
@@ -68,7 +68,7 @@ export const CovidVaccineCertificateModal = React.memo(({ open, onClose, patient
           printedDate={getCurrentDateString()}
           getLocalisation={getLocalisation}
         />
-      </PDFViewer>
+      </PDFLoader>
     </Modal>
   );
 });
