+++ packages/web/app/components/PatientPrinting/modals/EncounterRecordModal.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { PDFViewer } from '@react-pdf/renderer';
+import { PDFLoader } from '@react-pdf/renderer';
 
 import { NOTE_TYPES } from '@tamanu/constants/notes';
 import { LAB_REQUEST_STATUSES } from '@tamanu/constants/labs';
@@ -308,7 +308,7 @@ export const EncounterRecordModal = ({ encounter, open, onClose }) => {
 
   return (
     <Modal {...modalProps} onPrint={() => printPDF('encounter-record')}>
-      <PDFViewer
+      <PDFLoader
         isLoading={allQueries.isFetching}
         style={{ width: '100%', height: '600px' }}
         id="encounter-record"
@@ -334,7 +334,7 @@ export const EncounterRecordModal = ({ encounter, open, onClose }) => {
           getLocalisation={getLocalisation}
           clinicianText={clinicianText}
         />
-      </PDFViewer>
+      </PDFLoader>
     </Modal>
   );
 };
