+++ packages/web/app/components/PatientPrinting/modals/MultipleImagingRequestsPrintoutModal.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { PDFViewer, printPDF } from '../PDFViewer';
+import { PDFLoader, printPDF } from '../PDFLoader';
 import { MultipleImagingRequestsPrintout } from '@tamanu/shared/utils/patientCertificates';
 import { usePatientData } from '../../../api/queries/usePatientData';
 import { useReferenceData } from '../../../api/queries/useReferenceData';
@@ -18,7 +18,7 @@ export const MultipleImagingRequestsWrapper = ({ encounter, imagingRequests }) =
   const isLoading =
     isPatientLoading || (isVillageEnabled && isVillageLoading) || isCertificateFetching;
   return (
-    <PDFViewer isLoading={isLoading} id="imaging-request-printout">
+    <PDFLoader isLoading={isLoading} id="imaging-request-printout">
       <MultipleImagingRequestsPrintout
         getLocalisation={getLocalisation}
         patient={{ ...patient, village }}
@@ -26,7 +26,7 @@ export const MultipleImagingRequestsWrapper = ({ encounter, imagingRequests }) =
         imagingRequests={imagingRequests}
         certificateData={certificateData}
       />
-    </PDFViewer>
+    </PDFLoader>
   );
 };
 export const MultipleImagingRequestsPrintoutModal = ({
