+++ packages/web/app/components/PatientPrinting/modals/PatientStickerLabelPage.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useLocalisation } from '../../../contexts/Localisation';
-import { PDFViewer, printPDF } from '../PDFViewer';
+import { PDFLoader, printPDF } from '../PDFLoader';
 import { IDLabelPrintout } from '@tamanu/shared/utils/patientCertificates';
 import { Modal } from '../../Modal';
 
@@ -16,9 +16,9 @@ export const PatientStickerLabelPage = React.memo(({ patient }) => {
       printable
       onPrint={() => printPDF('patient-label-printout')}
     >
-      <PDFViewer id="patient-label-printout">
+      <PDFLoader id="patient-label-printout">
         <IDLabelPrintout patient={patient} measures={measures} />
-      </PDFViewer>
+      </PDFLoader>
     </Modal>
   );
 });
