+++ packages/web/app/views/patients/components/LabRequestPrintModal.jsx
@@ -10,7 +10,7 @@ import { useCertificate } from '../../../utils/useCertificate';
 
 import { Modal } from '../../../components';
 import { Colors } from '../../../constants';
-import { PDFViewer, printPDF } from '../../../components/PatientPrinting/PDFViewer';
+import { PDFLoader, printPDF } from '../../../components/PatientPrinting/PDFLoader';
 import { useLocalisation } from '../../../contexts/Localisation';
 import { useTranslation } from '../../../contexts/Translation';
 import { MultipleLabRequestsPrintout } from '@tamanu/shared/utils/patientCertificates';
@@ -62,7 +62,7 @@ export const LabRequestPrintModal = React.memo(({ labRequest, patient, open, onC
       printable
       onPrint={() => printPDF('lab-request-printout')}
     >
-      <PDFViewer isLoading={isLoading} id="lab-request-printout">
+      <PDFLoader isLoading={isLoading} id="lab-request-printout">
         <MultipleLabRequestsPrintout
           labRequests={[{ ...labRequest, tests: testsData.data, notes: notes?.data || [] }]}
           patientData={{ ...patient, additionalData, village }}
@@ -71,7 +71,7 @@ export const LabRequestPrintModal = React.memo(({ labRequest, patient, open, onC
           getLocalisation={getLocalisation}
           getTranslation={getTranslation}
         />
-      </PDFViewer>
+      </PDFLoader>
     </Modal>
   );
 });
