+++ packages/web/app/views/patients/imagingRequest/PrintModalButton.jsx
@@ -7,7 +7,7 @@ import { Modal } from '../../../components/Modal';
 import { TranslatedText } from '../../../components/Translation/TranslatedText';
 import { MultipleImagingRequestsWrapper } from '../../../components/PatientPrinting/modals/MultipleImagingRequestsPrintoutModal';
 import { Colors } from '../../../constants';
-import { printPDF } from '../../../components/PatientPrinting/PDFViewer';
+import { printPDF } from '../../../components/PatientPrinting/PDFLoader';
 
 const PrintModalInternals = ({ imagingRequest }) => {
   const encounterQuery = useEncounterData(imagingRequest.encounterId);
